README.md
Hello You are currently Reading the Official README File of the FNAF Mod on Friday Night Funkin'. 
This Project was started on 22nd July 2021 or In other words, 22/7/2021.
This is the Official Source code. 
If you are looking here to Understand the Code of whatever things like Custom Arrows, We Will leave Helpful Comments on the Hx files. 
This Mod is owned by 304em. 
This mod is coded by Renn. 
The music Composer for this mod is Local Darkner(The one who's writing this). 
If you are looking for changes on this mod, Please Credit The FNAF Team. 
Please do not make NSFW Out of this mod or any of its music [Lol how do u think of doing that? I MEAN NO SUS REMIXES. Erhmmm I mean sure] or arts.
BETADCIU's and Covers are Allowed. Join the Musician'S Discord: https://discord.gg/xdjnEugPrw
Join the Artist's Discord: https://discord.gg/5FpJEjVGBc
Not sure If the coder has a discord

Alright BYE

btw lol felt like putting these in
 __                 /                                  
/ _   _ |_ |_ .  _                                     
\__) (- |_ |_ | | )                                    
                                                       
 __                                                    
|_   _  _  _  |                                        
|   |  (- (_| |( \/                                    
                 /                                     
                                                       
 _   _     _                                           
(_) | )   (_|                                          
                                                       
 __                                                    
|_   _ .  _|  _        _  .  _  |_  |_       _  _  |_  
|   |  | (_| (_| \/   | ) | (_) | ) |_   \/ (- (_| | ) 
                 /          _/           /             




##################################################################################
                                                                                  
 _   _   _____   _      _    ____   ____     ___    _   _   _   _   ____    ____  
| \ | | | ____| | |    | |  / ___| |  _ \   / _ \  | | | | | \ | | |  _ \  / ___| 
|  \| | |  _|   | | /\ | | | |  _  | |_) | | | | | | | | | |  \| | | | | | \___ \ 
| |\  | | |___   \ V  V /  | |_| | |  _ <  | |_| | | |_| | | |\  | | |_| |  ___) |
|_| \_| |_____|   \_/\_/    \____| |_| \_\  \___/   \___/  |_| \_| |____/  |____/ 
 _       _  _     ___    ___       __              _       _  _       _        _  
|_ |  | |_ |_) \_/ | |_|  |  |\ | /__     |_      |_ |  | |_ |_) \_/ / \ |\ | |_  
|_  \/  |_ | \  |  | | | _|_ | \| \_| o   |_)\/   |_  \/  |_ | \  |  \_/ | \| |_ o
                                      /      /                                    
                                                                                  
##################################################################################